save progress added models to the stream reasponse
