--- extensions/toolhouse/tools.go
@@ -1,8 +1,0 @@
-package toolhouse
-
-const (
-	getToolsEndpoint = "/get_tools"
-)
-
-type Tool struct {
-}
