+++ unit_test.go
@@ -17,7 +17,6 @@ import (
 	"strconv"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/conneroisu/groq-go"
 	"github.com/conneroisu/groq-go/internal/test"
@@ -66,322 +65,6 @@ You have a six-sided die that you roll once. Let $R{i}$ denote the event that th
 	}
 }
 
-// TestCompletionsStreamWrongModel tests the completion stream returns an error when the model is not supported.
-func TestCompletionsStreamWrongModel(t *testing.T) {
-	a := assert.New(t)
-	client, err := groq.NewClient(
-		"whatever",
-		groq.WithBaseURL("http://localhost/v1"),
-	)
-	a.NoError(err, "NewClient returned error")
-
-	_, err = client.CreateCompletionStream(
-		context.Background(),
-		groq.CompletionRequest{
-			MaxTokens: 5,
-			Model:     groq.ModelWhisperLargeV3,
-		},
-	)
-	if !errors.Is(
-		err,
-		groq.ErrCompletionStreamNotSupported{},
-	) {
-		t.Fatalf(
-			"CreateCompletion should return ErrCompletionUnsupportedModel, but returned: %v",
-			err,
-		)
-	}
-}
-
-func TestCreateCompletionStream(t *testing.T) {
-	a := assert.New(t)
-	client, server, teardown := setupGroqTestServer()
-	defer teardown()
-	server.RegisterHandler(
-		"/v1/completions",
-		func(w http.ResponseWriter, _ *http.Request) {
-			w.Header().Set("Content-Type", "text/event-stream")
-
-			// Send test responses
-			dataBytes := []byte{}
-			dataBytes = append(dataBytes, []byte("event: message\n")...)
-			data := `{"id":"1","object":"completion","created":1598069254,"model":"text-davinci-002","choices":[{"text":"response1","finish_reason":"max_tokens"}]}`
-			dataBytes = append(dataBytes, []byte("data: "+data+"\n\n")...)
-
-			dataBytes = append(dataBytes, []byte("event: message\n")...)
-			data = `{"id":"2","object":"completion","created":1598069255,"model":"text-davinci-002","choices":[{"text":"response2","finish_reason":"max_tokens"}]}`
-			dataBytes = append(dataBytes, []byte("data: "+data+"\n\n")...)
-
-			dataBytes = append(dataBytes, []byte("event: done\n")...)
-			dataBytes = append(dataBytes, []byte("data: [DONE]\n\n")...)
-
-			_, err := w.Write(dataBytes)
-			a.NoError(err, "Write error")
-		},
-	)
-
-	stream, err := client.CreateCompletionStream(
-		context.Background(),
-		groq.CompletionRequest{
-			Prompt:    "Ex falso quodlibet",
-			Model:     "text-davinci-002",
-			MaxTokens: 10,
-			Stream:    true,
-		},
-	)
-	a.NoError(err, "CreateCompletionStream returned error")
-	defer stream.Close()
-
-	expectedResponses := []groq.CompletionResponse{
-		{
-			ID:      "1",
-			Object:  "completion",
-			Created: 1598069254,
-			Model:   "text-davinci-002",
-			Choices: []groq.CompletionChoice{
-				{Text: "response1", FinishReason: "max_tokens"},
-			},
-		},
-		{
-			ID:      "2",
-			Object:  "completion",
-			Created: 1598069255,
-			Model:   "text-davinci-002",
-			Choices: []groq.CompletionChoice{
-				{Text: "response2", FinishReason: "max_tokens"},
-			},
-		},
-	}
-
-	for ix, expectedResponse := range expectedResponses {
-		receivedResponse, streamErr := stream.Recv()
-		if streamErr != nil {
-			t.Errorf("stream.Recv() failed: %v", streamErr)
-		}
-		if !compareResponses(expectedResponse, receivedResponse) {
-			t.Errorf(
-				"Stream response %v is %v, expected %v",
-				ix,
-				receivedResponse,
-				expectedResponse,
-			)
-		}
-	}
-
-	_, streamErr := stream.Recv()
-	if !errors.Is(streamErr, io.EOF) {
-		t.Errorf("stream.Recv() did not return EOF in the end: %v", streamErr)
-	}
-
-	_, streamErr = stream.Recv()
-	if !errors.Is(streamErr, io.EOF) {
-		t.Errorf(
-			"stream.Recv() did not return EOF when the stream is finished: %v",
-			streamErr,
-		)
-	}
-}
-
-// TestCreateCompletionStreamTooManyEmptyStreamMessagesError tests the completion stream returns an error when the stream has too many empty messages.
-func TestCreateCompletionStreamTooManyEmptyStreamMessagesError(t *testing.T) {
-	a := assert.New(t)
-	client, server, teardown := setupGroqTestServer()
-	defer teardown()
-	server.RegisterHandler(
-		"/v1/completions",
-		func(w http.ResponseWriter, _ *http.Request) {
-			w.Header().Set("Content-Type", "text/event-stream")
-
-			// Send test responses
-			dataBytes := []byte{}
-			dataBytes = append(dataBytes, []byte("event: message\n")...)
-			data := `{"id":"1","object":"completion","created":1598069254,"model":"text-davinci-002","choices":[{"text":"response1","finish_reason":"max_tokens"}]}`
-			dataBytes = append(dataBytes, []byte("data: "+data+"\n\n")...)
-
-			// Totally 301 empty messages (300 is the limit)
-			for i := 0; i < 299; i++ {
-				dataBytes = append(dataBytes, '\n')
-			}
-
-			dataBytes = append(dataBytes, []byte("event: message\n")...)
-			data = `{"id":"2","object":"completion","created":1598069255,"model":"text-davinci-002","choices":[{"text":"response2","finish_reason":"max_tokens"}]}`
-			dataBytes = append(dataBytes, []byte("data: "+data+"\n\n")...)
-
-			dataBytes = append(dataBytes, []byte("event: done\n")...)
-			dataBytes = append(dataBytes, []byte("data: [DONE]\n\n")...)
-
-			_, err := w.Write(dataBytes)
-			a.NoError(err, "Write error")
-		},
-	)
-
-	stream, err := client.CreateCompletionStream(
-		context.Background(),
-		groq.CompletionRequest{
-			Prompt:    "Ex falso quodlibet",
-			Model:     "text-davinci-002",
-			MaxTokens: 10,
-			Stream:    true,
-		},
-	)
-	a.NoError(err, "CreateCompletionStream returned error")
-	defer stream.Close()
-
-	_, _ = stream.Recv()
-	_, streamErr := stream.Recv()
-	if !errors.Is(streamErr, groq.ErrTooManyEmptyStreamMessages{}) {
-		t.Errorf(
-			"TestCreateCompletionStreamTooManyEmptyStreamMessagesError did not return ErrTooManyEmptyStreamMessages",
-		)
-	}
-}
-
-// TestCreateCompletionStreamUnexpectedTerminatedError tests the completion stream returns an error when the stream is terminated unexpectedly.
-func TestCreateCompletionStreamUnexpectedTerminatedError(t *testing.T) {
-	a := assert.New(t)
-	client, server, teardown := setupGroqTestServer()
-	defer teardown()
-	server.RegisterHandler(
-		"/v1/completions",
-		func(w http.ResponseWriter, _ *http.Request) {
-			w.Header().Set("Content-Type", "text/event-stream")
-
-			// Send test responses
-			dataBytes := []byte{}
-			dataBytes = append(dataBytes, []byte("event: message\n")...)
-			data := `{"id":"1","object":"completion","created":1598069254,"model":"text-davinci-002","choices":[{"text":"response1","finish_reason":"max_tokens"}]}`
-			dataBytes = append(dataBytes, []byte("data: "+data+"\n\n")...)
-
-			// Stream is terminated without sending "done" message
-
-			_, err := w.Write(dataBytes)
-			a.NoError(err, "Write error")
-		},
-	)
-
-	stream, err := client.CreateCompletionStream(
-		context.Background(),
-		groq.CompletionRequest{
-			Prompt:    "Ex falso quodlibet",
-			Model:     "text-davinci-002",
-			MaxTokens: 10,
-			Stream:    true,
-		},
-	)
-	a.NoError(err, "CreateCompletionStream returned error")
-	defer stream.Close()
-
-	_, _ = stream.Recv()
-	_, streamErr := stream.Recv()
-	if !errors.Is(streamErr, io.EOF) {
-		t.Errorf(
-			"TestCreateCompletionStreamUnexpectedTerminatedError did not return io.EOF",
-		)
-	}
-}
-
-// TestCreateCompletionStreamBrokenJSONError tests the completion stream returns an error when the stream is broken.
-func TestCreateCompletionStreamBrokenJSONError(t *testing.T) {
-	a := assert.New(t)
-	client, server, teardown := setupGroqTestServer()
-	defer teardown()
-	server.RegisterHandler(
-		"/v1/completions",
-		func(w http.ResponseWriter, _ *http.Request) {
-			w.Header().Set("Content-Type", "text/event-stream")
-
-			// Send test responses
-			dataBytes := []byte{}
-			dataBytes = append(dataBytes, []byte("event: message\n")...)
-			data := `{"id":"1","object":"completion","created":1598069254,"model":"text-davinci-002","choices":[{"text":"response1","finish_reason":"max_tokens"}]}`
-			dataBytes = append(dataBytes, []byte("data: "+data+"\n\n")...)
-
-			// Send broken json
-			dataBytes = append(dataBytes, []byte("event: message\n")...)
-			data = `{"id":"2","object":"completion","created":1598069255,"model":`
-			dataBytes = append(dataBytes, []byte("data: "+data+"\n\n")...)
-
-			dataBytes = append(dataBytes, []byte("event: done\n")...)
-			dataBytes = append(dataBytes, []byte("data: [DONE]\n\n")...)
-
-			_, err := w.Write(dataBytes)
-			a.NoError(err, "Write error")
-		},
-	)
-
-	stream, err := client.CreateCompletionStream(
-		context.Background(),
-		groq.CompletionRequest{
-			Prompt:    "Ex falso quodlibet",
-			Model:     "text-davinci-002",
-			MaxTokens: 10,
-			Stream:    true,
-		},
-	)
-	a.NoError(err, "CreateCompletionStream returned error")
-	defer stream.Close()
-
-	_, _ = stream.Recv()
-	_, streamErr := stream.Recv()
-	var syntaxError *json.SyntaxError
-	if !errors.As(streamErr, &syntaxError) {
-		t.Errorf(
-			"TestCreateCompletionStreamBrokenJSONError did not return json.SyntaxError",
-		)
-	}
-}
-
-func TestCreateCompletionStreamReturnTimeoutError(t *testing.T) {
-	client, server, teardown := setupGroqTestServer()
-	defer teardown()
-	server.RegisterHandler(
-		"/v1/completions",
-		func(http.ResponseWriter, *http.Request) {
-			time.Sleep(10 * time.Nanosecond)
-		},
-	)
-	ctx := context.Background()
-	ctx, cancel := context.WithTimeout(ctx, time.Nanosecond)
-	defer cancel()
-
-	_, err := client.CreateCompletionStream(ctx, groq.CompletionRequest{
-		Prompt:    "Ex falso quodlibet",
-		Model:     "text-davinci-002",
-		MaxTokens: 10,
-		Stream:    true,
-	})
-	if err == nil {
-		t.Fatal("Did not return error")
-	}
-	if !os.IsTimeout(err) {
-		t.Fatal("Did not return timeout error")
-	}
-}
-
-// Helper funcs.
-func compareResponses(r1, r2 groq.CompletionResponse) bool {
-	if r1.ID != r2.ID || r1.Object != r2.Object || r1.Created != r2.Created ||
-		r1.Model != r2.Model {
-		return false
-	}
-	if len(r1.Choices) != len(r2.Choices) {
-		return false
-	}
-	for i := range r1.Choices {
-		if !compareResponseChoices(r1.Choices[i], r2.Choices[i]) {
-			return false
-		}
-	}
-	return true
-}
-
-func compareResponseChoices(c1, c2 groq.CompletionChoice) bool {
-	if c1.Text != c2.Text || c1.FinishReason != c2.FinishReason {
-		return false
-	}
-	return true
-}
-
 // TestModerate tests the Moderate method of the client.
 func TestModerate(t *testing.T) {
 	client, server, teardown := setupGroqTestServer()
@@ -391,7 +74,7 @@ func TestModerate(t *testing.T) {
 		handleModerationEndpoint,
 	)
 	mod, err := client.Moderate(context.Background(), groq.ModerationRequest{
-		Model: groq.ModerationTextStable,
+		Model: groq.ModelLlamaGuard38B,
 		Messages: []groq.ChatCompletionMessage{
 			{
 				Role:    groq.ChatMessageRoleUser,
@@ -417,7 +100,7 @@ func handleModerationEndpoint(w http.ResponseWriter, r *http.Request) {
 		ID:      "chatcmpl-123",
 		Object:  "chat.completion",
 		Created: 1693721698,
-		Model:   groq.ModerationTextStable,
+		Model:   groq.ModelLlamaGuard38B,
 		Choices: []groq.ChatCompletionChoice{
 			{
 				Message: groq.ChatCompletionMessage{
@@ -449,91 +132,7 @@ func handleModerationEndpoint(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-// TestCompletionWithStream tests the CreateCompletion method with a stream.
-func TestCompletionWithStream(t *testing.T) {
-	a := assert.New(t)
-	client, err := groq.NewClient(
-		"whatever",
-		groq.WithBaseURL("http://localhost/v1"),
-	)
-	a.NoError(err, "NewClient error")
-
-	ctx := context.Background()
-	req := groq.CompletionRequest{Stream: true}
-	_, err = client.CreateCompletion(ctx, req)
-	if !errors.Is(err, groq.ErrCompletionStreamNotSupported{}) {
-		t.Fatalf(
-			"CreateCompletion didn't return ErrCompletionStreamNotSupported",
-		)
-	}
-}
-
-// TestCompletions Tests the completions endpoint of the API using the mocked server.
-func TestCompletions(t *testing.T) {
-	a := assert.New(t)
-	client, server, teardown := setupGroqTestServer()
-	defer teardown()
-	server.RegisterHandler("/v1/completions", handleCompletionEndpoint)
-	req := groq.CompletionRequest{
-		MaxTokens: 5,
-		Model:     "ada",
-		Prompt:    "Lorem ipsum",
-	}
-	_, err := client.CreateCompletion(context.Background(), req)
-	a.NoError(err, "CreateCompletion error")
-}
-
-// handleCompletionEndpoint Handles the completion endpoint by the test server.
-func handleCompletionEndpoint(w http.ResponseWriter, r *http.Request) {
-	var err error
-	var resBytes []byte
-
-	// completions only accepts POST requests
-	if r.Method != "POST" {
-		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
-	}
-	var completionReq groq.CompletionRequest
-	if completionReq, err = getCompletionBody(r); err != nil {
-		http.Error(w, "could not read request", http.StatusInternalServerError)
-		return
-	}
-	res := groq.CompletionResponse{
-		ID:      strconv.Itoa(int(time.Now().Unix())),
-		Object:  "test-object",
-		Created: time.Now().Unix(),
-		// would be nice to validate Model during testing, but
-		// this may not be possible with how much upkeep
-		// would be required / wouldn't make much sense
-		Model: completionReq.Model,
-	}
-	// create completions
-	n := completionReq.N
-	if n == 0 {
-		n = 1
-	}
-	for i := 0; i < n; i++ {
-		// generate a random string of length completionReq.Length
-		completionStr := strings.Repeat("a", completionReq.MaxTokens)
-		if completionReq.Echo {
-			completionStr = completionReq.Prompt.(string) + completionStr
-		}
-		res.Choices = append(res.Choices, groq.CompletionChoice{
-			Text:  completionStr,
-			Index: i,
-		})
-	}
-	inputTokens := numTokens(completionReq.Prompt.(string)) * n
-	completionTokens := completionReq.MaxTokens * n
-	res.Usage = groq.Usage{
-		PromptTokens:     inputTokens,
-		CompletionTokens: completionTokens,
-		TotalTokens:      inputTokens + completionTokens,
-	}
-	resBytes, _ = json.Marshal(res)
-	fmt.Fprintln(w, string(resBytes))
-}
-
-// numTokens Returns the number of GPT-3 encoded tokens in the given text.
+// numTokens Returns the number of encoded tokens in the given text.
 // This function approximates based on the rule of thumb stated by OpenAI:
 // https://beta.openai.com/tokenizer
 //
@@ -541,21 +140,6 @@ func handleCompletionEndpoint(w http.ResponseWriter, r *http.Request) {
 // instead.
 func numTokens(s string) int {
 	return int(float32(len(s)) / 4)
-}
-
-// getCompletionBody Returns the body of the request to create a completion.
-func getCompletionBody(r *http.Request) (groq.CompletionRequest, error) {
-	completion := groq.CompletionRequest{}
-	// read the request body
-	reqBody, err := io.ReadAll(r.Body)
-	if err != nil {
-		return groq.CompletionRequest{}, err
-	}
-	err = json.Unmarshal(reqBody, &completion)
-	if err != nil {
-		return groq.CompletionRequest{}, err
-	}
-	return completion, nil
 }
 
 func setupGroqTestServer() (
